refactor: Color와 Type enum 분리
