+++ src/main/java/com/woopaca/javachess/chess/Rank.java
@@ -1,8 +1,8 @@
 package com.woopaca.javachess.chess;
 
+import com.woopaca.javachess.pieces.Color;
 import com.woopaca.javachess.pieces.Piece;
-import com.woopaca.javachess.pieces.Piece.Color;
-import com.woopaca.javachess.pieces.Piece.Type;
+import com.woopaca.javachess.pieces.Type;
 
 import java.util.ArrayList;
 import java.util.List;
