+++ src/main/java/com/woopaca/javachess/pieces/Color.java
@@ -1,0 +1,7 @@
+package com.woopaca.javachess.pieces;
+
+public enum Color {
+
+    WHITE, BLACK, NOCOLOR
+
+}
